Fix sudo usage

+++ .jenkins.d/01-ndn-cxx.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 set -ex
 
-sudo rm -f /usr/local/bin/ndnsec*
-sudo rm -fr /usr/local/include/ndn-cxx
-sudo rm -f /usr/local/lib{,64}/libndn-cxx*
-sudo rm -f /usr/local/lib{,64}/pkgconfig/libndn-cxx.pc
+if ! has Windows $NODE_LABELS; then
+    sudo rm -f /usr/local/bin/ndnsec*
+    sudo rm -fr /usr/local/include/ndn-cxx
+    sudo rm -f /usr/local/lib{,64}/libndn-cxx*
+    sudo rm -f /usr/local/lib{,64}/pkgconfig/libndn-cxx.pc
+fi
